pyproject.toml
[tool.poetry]
name = "pyrevolut"
version = "0.5.1"
description = ""
authors = ["Trevor Visser <trevor.visser@gmail.com>"]
readme = "README.md"

[tool.poetry.scripts]
pyrevolut = "pyrevolut.cli.main:app"

[tool.poetry.dependencies]
python = "^3.11"
httpx = "^0.27.0"
pydantic = { extras = ["email"], version = "^2.7.1" }
pydantic-extra-types = "^2.7.0"
pycountry = "^23.12.11"
phonenumbers = "^8.13.36"
pendulum = "^3.0.0"
pyjwt = { extras = ["crypto"], version = "^2.8.0" }
cryptography = "^42.0.7"
authlib = "^1.3.0"
typer = "^0.12.3"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.0"
ipykernel = "^6.29.4"
commitizen = "^3.25.0"

[tool.poetry.group.test.dependencies]
pytest-asyncio = "^0.21.0"
pytest = "^7.3.1"
pytest-xdist = { extras = ["psutil"], version = "^3.3.1" }
pytest-cov = "^4.1.0"
ruff = "*"
black = "^24.4.2"
pytest-sugar = "^0.9.7"
pytest-instafail = "^0.5.0"
pytest-profiling = "^1.7.0"
pytest-memray = "^1.5.0"
pytest-randomly = "^3.15.0"
pytest-clarity = "^1.0.1"
pytest-network = "^0.0.1"
pytest-resource-usage = "^1.0.0"
pytest-split = "^0.8.1"
pytest-env = "^1.1.3"
locust = "^2.18.1"
requests-toolbelt = "^1.0.0"

[tool.pytest.ini_options]
pythonpath = ["."]
filterwarnings = []
addopts = "--instafail -vv"
env = ["TEST_MODE=True"]
asyncio_mode = "auto"
log_cli = false
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"

[tool.coverage.report]
show_missing = true
exclude_lines = [
    # Have to re-enable the standard pragma
    'pragma: no cover',

    # Don't complain about missing debug-only code:
    'def __repr__',
    'if self\.debug',

    # Don't complain if tests don't hit defensive assertion code:
    'raise AssertionError',
    'raise NotImplementedError',

    # Don't complain if non-runnable code isn't run:
    'if 0:',
    'if __name__ == .__main__.:',
    'if TYPE_CHECKING:',
]
omit = [
    'env/*',
    'venv/*',
    '.venv/*',
    '*/virtualenv/*',
    '*/virtualenvs/*',
    '*/tests/*',
]

[tool.coverage.run]
omit = [".*", "*/site-packages/*", "tests/*"]

[tool.coverage.json]
pretty_print = true

[tool.mypy]
files = "pyrevolut"
disallow_untyped_decorators = true
disallow_subclassing_any = true
warn_return_any = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_unused_configs = true

[tool.ruff]
line-length = 88

[tool.ruff.format]
quote-style = "double"
docstring-code-format = true

[tool.ruff.lint.isort]
known-first-party = ["pyrevolut", "tests"]

[tool.commitizen]
name = "cz_conventional_commits"
version = "0.5.1"
tag_format = "v$major.$minor.$patch$prerelease"
version_scheme = "pep440"
version_provider = "poetry"
update_changelog_on_bump = true
major_version_zero = true
version_files = ["pyrevolut/__init__.py:__version__", "pyproject.toml:^version"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
